+++ csvviz/cmds/area.py
@@ -31,7 +31,7 @@ class Areakit(Vizkit):
             help="The name of the column for mapping y-axis values; if empty, the second (columns[1]) column is used",
         ),
         click.option(
-            "--color",
+            "--colorvar",
             "-c",
             "fillvar",
             type=click.STRING,
@@ -45,18 +45,34 @@ class Areakit(Vizkit):
             type=click.Choice(("asc", "desc"), case_sensitive=False),
             help="For stacked charts, the sort order of the color variable: 'asc' for ascending, 'desc' for descending/reverse",
         ),
+        click.option(
+            "--normalized",
+            "-N",
+            is_flag=True,
+            help="For stacked bar charts, normalize the total area heights to 100%",
+        ),
     )
 
     def prepare_channels(self):
         channels = self._create_channels(self.channel_kwargs)
         self._set_channel_colorscale("fill", channels)
 
+        # https://altair-viz.github.io/gallery/normalized_stacked_area_chart.html
+        if self.kwargs.get("normalized"):
+            if not channels.get("fill"):
+                raise MissingDataReference(
+                    "-c/--colorvar needs to be specified when creating a normalized (i.e. stacked) chart"
+                )
+            else:
+                channels["y"].stack = "normalize"
+                channels["y"].axis = alt.Axis(format="%")
+
         # sort by fill/stack is not the same as sorting the x-axis:
         # https://altair-viz.github.io/user_guide/encoding.html?#ordering-marks
         if _fillsort := self.kwargs.get("fillsort"):
             if not channels.get("fill"):
                 raise MissingDataReference(
-                    f"--color-sort '{_fillsort}' was specified, but no --color value was provided"
+                    f"--color-sort '{_fillsort}' was specified, but no --colorvar value was provided"
                 )
             else:
                 fname = self.resolve_channel_name(channels["fill"])