+++ tests/unit/test_legends.py
@@ -27,7 +27,7 @@ DEFAULT_ARGS = [
 # legend
 ##############################################################################################################
 def test_legend_default():
-    """when there is a fill --color, there is a legend"""
+    """when there is a fill --colorvar, there is a legend"""
     result = CliRunner().invoke(viz, ["-c", "breed", *DEFAULT_ARGS])
     cdata = json.loads(result.output)
     legend = cdata["encoding"]["fill"]["legend"]
@@ -40,7 +40,8 @@ def test_no_legend():
     """hiding the legend sets fill.legend to None explicitly"""
 
     result = CliRunner().invoke(
-        viz, ["--color", "breed", "--size", "velocity", "--no-legend", *DEFAULT_ARGS]
+        viz,
+        ["--colorvar", "breed", "--sizevar", "velocity", "--no-legend", *DEFAULT_ARGS],
     )
     cdata = json.loads(result.output)